+++ Makefile
@@ -4,18 +4,18 @@ help:
 	@awk 'BEGIN {FS = ":.*##"; printf "Usage: make \033[36m<target>\033[0m\n"} /^[a-zA-Z_-]+:.*?##/ { printf "  \033[36m%-10s\033[0m %s\n", $$1, $$2 } /^##@/ { printf "\n\033[1m%s\033[0m\n", substr($$0, 5) } ' $(MAKEFILE_LIST)
 
 run:
-	docker-compose -f config/dev/docker-compose.yml up -d web-app db
+	docker-compose -f conf/dev/docker-compose.yml up -d web-app db redis
 
 stop:
-	docker-compose -f config/dev/docker-compose.yml down -v --remove-orphans
+	docker-compose -f conf/dev/docker-compose.yml down -v --remove-orphans
 
 rebuild_app:
 	docker image rm dev-web-app
-	docker-compose -f config/dev/docker-compose.yml build
+	docker-compose -f conf/dev/docker-compose.yml build
 
 rebuild_test:
 	docker image rm dev-test
-	docker-compose -f config/dev/docker-compose.yml build
+	docker-compose -f conf/dev/docker-compose.yml build
 
 delete_postgres_volume:
 	docker volume rm dev_menu-db
@@ -23,6 +23,10 @@ delete_postgres_volume:
 restart_app:
 	docker stop dev-web-app-1
 	docker start dev-web-app-1
+
+restart_redis:
+	docker stop dev-redis-1
+	docker start dev-redis-1
 
 logs_app:
 	docker logs dev-web-app-1
@@ -40,4 +44,4 @@ alembic_migrate:
 	docker exec -ti dev-web-app-1 alembic upgrade head
 
 test_run:
-	docker-compose -f config/dev/docker-compose.yml up test
+	docker-compose -f conf/dev/docker-compose.yml up test
